added glm and logistic regression
